# Don't return to "readBlock" when "Try" fails.